+++ src/components/Map/FocusedLocationBottomSheet/FocusedLocationBottomSheet.module.css
@@ -1,4 +1,4 @@
-.LocationInfoBottomSheetContainer {
+.DetailInfoBottomSheetContainer {
   position: relative;
   width: 100%;
   position: absolute;
@@ -14,27 +14,11 @@
   opacity: 0;
   transition: transform 0.3s ease-in-out;
 }
-.LocationInfoBottomSheetContainer.open {
+.DetailInfoBottomSheetContainer.open {
   transform: translateY(0%);
   opacity: 1;
 }
-.ListButton {
-  padding: 9px 13px;
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  border-radius: 18px;
-  border: 2px solid var(--Gray-400, #e3e7eb);
-  background: #fff;
-}
-.ListButtonText {
-  color: var(--Black, #1d2228);
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 120%; /* 18px */
-}
-.LocationInfoBottomSheet {
+.DetailInfoBottomSheet {
   height: 380px;
   width: 100%;
   max-width: 700px;
@@ -63,7 +47,7 @@
   top: 10px;
   z-index: 10; /* 다른 요소 위에 오도록 */
 }
-.LocationInfoWrapper {
+.DetailInfoWrapper {
   width: calc(100% - 28px);
   padding: 20px 0;
   box-sizing: border-box;
@@ -102,15 +86,6 @@
   flex-direction: column;
   gap: 10px;
 }
-.BottomSheetGrad {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 30px;
-  background: linear-gradient(180deg, rgba(255, 255, 255, 0) 0%, #fff 66.83%);
-  z-index: 1000;
-}
-
 .TitleText,
 .SubTitleText,
 .FriendTitle,
@@ -138,4 +113,11 @@
   font-size: 16px;
   font-weight: 500;
   line-height: 120%; /* 19.2px */
+}
+
+.ButtonContainer {
+  position: fixed;
+  padding: 15px 27px;
+  width: 100%;
+  box-sizing: border-box;
 }
